renamed common to generic to fit with current rust bridge naming
