+++ Rust/hashcat-sys/src/generic.rs
